double

+++ ContaEspecial.java
@@ -1,10 +1,10 @@
 public class ContaEspecial extends Conta {
-  float limite;
-  public ContaEspecial(String nome, String endereco, int cpf, int data, float saldo, float limite){
+  double limite;
+  public ContaEspecial(String nome, String endereco, int cpf, int data, double saldo, double limite){
     super(nome, endereco, cpf, data, saldo);
     this.limite = limite;
   }
-  public boolean sacar(float valor){
+  public boolean sacar(double valor){
     if(valor > (super.saldo + limite)){
       return false;
     }
